# openapi.yaml
openapi: 3.1.0
info:
  title: Memoria API
  version: 1.0.0
  description: Private photo/video sharing API (Capsules + Albums). All times UTC in JSON.
servers:
  - url: http://localhost:8080
    description: Local docker-compose
  - url: https://api.memoria.app
    description: Production (placeholder)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
paths:
...
  /v1/me/stats/heatmap:
    get:
      tags: [stats]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Daily contribution counts }
  /v1/me/stats/lifetime:
    get:
      tags: [stats]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Lifetime totals }
...
  /v1/memories/{id}/reactions:
    put:
      tags: [social]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [emoji]
              properties:
                emoji: { type: string, maxLength: 32 }
      responses:
        '200': { description: Reaction upserted }
    delete:
      tags: [social]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '204': { description: Reaction removed }
  /v1/memories/{id}/comments:
    get:
      tags: [social]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/id'
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200': { description: Comments page }
    post:
      tags: [social]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, maxLength: 1000 }
      responses:
        '201': { description: Comment created }
  /v1/memories/{id}/comments/{commentId}:
    delete:
      tags: [social]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/id'
        - name: commentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Comment deleted }
...